Move test to AndroidSyncFrameworkTest
